+++ src/main/java/com/surit/fixer/job/model/mapper/JobMapper.java
@@ -15,11 +15,11 @@ public interface JobMapper {
 	 *
 	 * @param statusCode 상태 필터 (없으면 null → 전체)
 	 */
-	List<JobDTO> selectMyJobs(@Param("fixerNo") int fixerNo,
+	List<JobDTO> selectMyJobs(@Param("fixerNo") long fixerNo,
 	                          @Param("statusCode") String statusCode);
 
 	/** 내 작업 상세 (내 작업이 아니면 null) */
-	JobDTO selectMyJob(@Param("fixerNo") int fixerNo,
+	JobDTO selectMyJob(@Param("fixerNo") long fixerNo,
 	                   @Param("requestId") long requestId);
 
 	/**
@@ -31,6 +31,6 @@ public interface JobMapper {
 	 *
 	 * @return 1 이면 성공, 0 이면 조건 불일치
 	 */
-	int completeJob(@Param("fixerNo") int fixerNo,
+	int completeJob(@Param("fixerNo") long fixerNo,
 	                @Param("requestId") long requestId);
 }
